pyproject.toml
[tool.poetry]
name = "steerable_scene_generation"
version = "0.1.0"
description = ""
authors = ["Nicholas Pfaff <nepfaff@mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
torch = "^2.3.0"
diffusers = "^0.33.1"
tqdm = "^4.66.4"
numpy = ">=1.26.4,<3.0.0"
drake = "^1.28.0"
wandb = "^0.17.0"
hydra-core = "^1.3.2"
ipython = "^8.24.0"
pytorch3d = {url = "https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py310_cu118_pyt201/pytorch3d-0.7.4-cp310-cp310-linux_x86_64.whl"}
colorama = "^0.4.6"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.4"
lightning = "^2.2.4"
click = "^8.1.7"
imageio = "^2.34.1"
einops = "^0.8.0"
wandb-osh = "^1.2.2"
torchvision = "^0.18.0"
pyyaml = "^6.0.1"
trimesh = "<4.2.0"
pyvirtualdisplay = "^3.0"
roma = "^1.5.0"
clean-fid = "^0.1.35"
cmmd-pytorch = {git = "https://github.com/nepfaff/cmmd-pytorch.git", rev = "0cc906ea710ba900a7caf572e85fd3ca9fa2254f"}
prdc = "^0.2"
top-pr = "^0.2.1"
datasets = "^3.1.0"
sentencepiece = "^0.2.0"
openai = "^1.59.9"
graphviz = "^0.20.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "23.9.1"
isort = "5.12.0"
poetry-pre-commit-plugin = "^0.1.2"

[tool.isort]
profile = 'black'
lines_between_types = 1
combine_as_imports = true
known_first_party = ['steerable_scene_generation']
known_third_party = ['wandb']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
